[Oedax] fix minor issue (#196)

+++ packages/oedax_v1/contracts/iface/IOedax.sol
@@ -61,7 +61,7 @@ contract IOedax is Ownable
     uint16      public protocolFeeBips;
     uint16      public ownerFeeBips;
     uint16      public takerFeeBips;
-    uint16      public minDuration;
+    uint32      public minDuration;
     uint32      public maxDuration;
     uint        public creatorEtherStake;
     address[]   public auctions;
